revert to shared firstApp model on welcome route
