+++ app/welcome/payment-info/template.hbs
@@ -59,8 +59,8 @@
     </div>
     <div class="col-xs-5">
       <div class="welcome-info">
-        {{#if model.appHandle}}
-          <h2>Add a payment method to provision {{model.appHandle}}.</h2>
+        {{#if firstApp.appHandle}}
+          <h2>Add a payment method to provision {{firstApp.appHandle}}.</h2>
         {{else}}
           <h2>Add a payment method to your free Development account</h2>
         {{/if}}
